requirements.txt
Flask==3.0.0
pymongo>=4.15.0
python-dotenv==1.0.0
openai>=2.3.0
werkzeug==3.0.1
bcrypt==4.1.1
pandas>=2.1.3
